requirements.txt
celery==5.2.7
dydx-v3-python==1.9.0
matplotlib==3.6.2
pandas==1.5.1
pytest==4.6.11
python-binance==1.0.16
python-dotenv==0.21.0
stellar-sdk==8.1.1
